Fix behaviors
